Move the Control.DragDrop tests under ControlTests.Methods.cs to Collection "Sequential"
